chore: restore game modules and cleanup
