--- src/themes/clubs.ts
@@ -1,14 +1,0 @@
-export const CLUB_THEMES = {
-  liverpool: {
-    primary: 'bg-[#c8102e]', // Liverpool red
-    accent: 'bg-[#00b2a9]', // Liverpool teal
-    text: 'text-white',
-    logo: '/assets/liverpool.png',
-  },
-  madrid: {
-    primary: 'bg-[#ffffff]', // Madrid white
-    accent: 'bg-[#febd2f]', // Madrid gold
-    text: 'text-black',
-    logo: '/assets/realmadrid.png',
-  },
-};
